+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog].
   default is t) ([#261]).
 
 ### Enhancements
+* In file completions the prompt will also be selected when a match is
+  required and the path exists ([#357]).
 * With commands `next-history-element` and `previous-history-element`
   the inserted history element will get selected when a match isn't
   required which helps when the element isn't a member of the
@@ -219,6 +221,7 @@ The format is based on [Keep a Changelog].
 [#352]: https://github.com/raxod502/selectrum/pull/352
 [#354]: https://github.com/raxod502/selectrum/pull/354
 [#356]: https://github.com/raxod502/selectrum/pull/356
+[#357]: https://github.com/raxod502/selectrum/pull/357
 
 ## 3.0 (released 2020-10-20)
 ### Breaking changes